Add files via upload
This R code generates a tile plot using ggplot2 to visualize the number of attacks by weapon and region. 
Adjusting the size of the plot using options(repr.plot.width = 12, repr.plot.height = 12). This sets the plot width and height to 12 inches.
Creating the Tile Plot:
Using ggplot to create a tile plot (geom_tile) with regions on the x-axis (reorder(region, -count)) and weapons on the y-axis (reorder(weapon, count)).
The fill color of each tile represents the logarithm of the count of attacks (fill = logcount).
Labels are added to each tile (geom_text(size = 6)).
The color of the tile borders is set to white (color = "white"), and the size is set to 0.1.
The title of the plot is set to "Number of attacks by weapon."

This code generates a visually appealing tile plot that provides insights into the distribution of attacks across different regions and weapons. The use of logarithmic scaling for the fill color enhances the visibility of patterns in regions with varying attack counts. The resulting plot is saved as an image file for further use or sharing.